refactor(start): start.js refactor
